resolve re-rendering error by removing whitespaces in table row, hide nested objects from node-table, restructure table folder
