+++ src/js/engagement_view/src/components/engagementView/sidebar/LensAndNodeTableContainer.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 
-import NodeTable from '../nodeTable/NodeTable';
+import NodeTable from './tables/nodeTable/NodeTable';
 import { NodeDetailsProps, EngagementViewProps } from "types/LensAndNodeTableTypes";
 
-import {ToggleLensTable} from "./tableStructure/toggleLensTable";
-import {ToggleNodeDetailTable} from './tableStructure/toggleNodeDetailTable';
+import {ToggleLensTable} from "./tables/toggleLensTable";
+import {ToggleNodeDetailTable} from './tables/toggleNodeDetailTable';
 
 export const NodeDetails = ({node}: NodeDetailsProps) => {
     return (
